Update button help patch
Hello world example: sends T instead of hello world, as is the default value + I thought possibly confusing that the button text and send value are the same text.

Collect example:
- Removed the lambda, replaced by a simple call to om-random
- Added an OR including the Reset action in push
- Use a value box rather than a print

tcollect example: Removed. This is more an example of tcollect than button.
Also, the button actually doesn't react so well to multiple-clicks, so better not to emphasize this :)

Other minor layout changes.